requirements.txt
babel
python-dateutil==2.6.0
flask-moment
flask-wtf
flask_sqlalchemy==2.4.1
psycopg2-binary==2.8.4
flask-migrate==2.5.2
Flask-Script==2.0.6